requirements.txt
numpy==1.24.2
matplotlib==3.6.3
scipy==1.10.0
requests==2.28.2
google-auth-oauthlib==0.8.0
google-api-python-client==2.76.0